Changed double variables to float for division

+++ src/CelsiusConverter.java
@@ -1,6 +1,6 @@
 public class CelsiusConverter implements TemperatureConverter{
     public CelsiusConverter(){}
-    public double convertToCelsius(double amount){return amount;}
-    public double convertToFahrenheit(double amount){return ((amount * (9/5)) + 32); }
-    public double convertToKelvin(double amount){return amount + 273.15;}
+    public float convertToCelsius(float amount){return amount;}
+    public float convertToFahrenheit(float amount){return ((amount * (9f/5f)) + 32f); }
+    public float convertToKelvin(float amount){return amount + 273.15f;}
 }
